requirements.txt
pandas
pydrive
google-cloud-bigquery
pyarrow
pandas-gbq
python-dotenv
